Improve copying price strings for bulk items
When using the evaluate module and clicking on another
person's price listing to copy the price, if the price
is a fraction it currently copies as a decimal number, ie:
0.125.  However if you list it as such, listing services will
round it to 0.5.  Further, depending on the fraction, it
can be hard to reason about as a user on how many that
really represents per unit of currency.

This patch modifies the copied text to instead contain
the pervasive fractional numerator/denominator formatting for
any price that has a denominator greater than 1.